Enum move_ir_types::ast::Exp_

pub enum Exp_ {
    Dereference(Box<Exp>),
    UnaryExp(UnaryOp, Box<Exp>),
    BinopExp(Box<Exp>, BinOp, Box<Exp>),
    Value(CopyableVal),
    Pack(StructName, Vec<Type>, ExpFields),
    Borrow {
        is_mutable: bool,
        exp: Box<Exp>,
        field: Field_,
    },
    Move(Var),
    Copy(Var),
    BorrowLocal(bool, Var),
    FunctionCall(FunctionCall, Box<Exp>),
    ExprList(Vec<Exp>),
}

Enum for all expressions

Variants

Dereference(Box<Exp>)

*e

UnaryExp(UnaryOp, Box<Exp>)

op e

BinopExp(Box<Exp>, BinOp, Box<Exp>)

e_1 op e_2

Value(CopyableVal)

Wrapper to lift CopyableVal into Exp v

Pack(StructName, Vec<Type>, ExpFields)

Takes the given field values and instantiates the struct Returns a fresh StructInstance whose type and kind (resource or otherwise) as the current struct class (i.e., the class of the method we’re currently executing). n { f_1: e_1, ... , f_j: e_j }

Borrow

&e.f, &mut e.f

Fields of Borrow

is_mutable: bool

mutable or not

exp: Box<Exp>

the expression containing the reference

field: Field_

the field being borrowed

Move(Var)

move(x)

Copy(Var)

copy(x)

BorrowLocal(bool, Var)

&x or &mut x

FunctionCall(FunctionCall, Box<Exp>)

f(e) or f(e_1, e_2, ..., e_j)

ExprList(Vec<Exp>)

(e_1, e_2, e_3, …, e_j)

Implementations

impl Exp_

pub fn address(addr: AccountAddress) -> Exp

Creates a new address Exp with no location information

pub fn value(b: CopyableVal_) -> Exp

Creates a new value Exp with no location information

pub fn u64(i: u64) -> Exp

Creates a new u64 Exp with no location information

pub fn bool(b: bool) -> Exp

Creates a new bool Exp with no location information

pub fn byte_array(buf: Vec<u8>) -> Exp

Creates a new bytearray Exp with no location information

pub fn instantiate(n: StructName, tys: Vec<Type>, s: ExpFields) -> Exp

Creates a new pack/struct-instantiation Exp with no location information

pub fn binop(lhs: Exp, op: BinOp, rhs: Exp) -> Exp

Creates a new binary operator Exp with no location information

pub fn add(lhs: Exp, rhs: Exp) -> Exp

Creates a new e+e Exp with no location information

pub fn sub(lhs: Exp, rhs: Exp) -> Exp

Creates a new e-e Exp with no location information

pub fn dereference(e: Exp) -> Exp

Creates a new *e Exp with no location information

pub fn borrow(is_mutable: bool, exp: Box<Exp>, field: Field_) -> Exp

Creates a new borrow field Exp with no location information

pub fn copy(v: Var) -> Exp

Creates a new copy-local Exp with no location information

pub fn move_(v: Var) -> Exp

Creates a new move-local Exp with no location information

pub fn function_call(f: FunctionCall, e: Exp) -> Exp

Creates a new function call Exp with no location information

pub fn expr_list(exps: Vec<Exp>) -> Exp