Struct move_ir_types::ast::ModuleDefinition

pub struct ModuleDefinition {
    pub loc: Loc,
    pub identifier: ModuleIdent,
    pub friends: Vec<ModuleIdent>,
    pub imports: Vec<ImportDefinition>,
    pub explicit_dependency_declarations: Vec<ModuleDependency>,
    pub structs: Vec<StructDefinition>,
    pub constants: Vec<Constant>,
    pub functions: Vec<(FunctionName, Function)>,
    pub synthetics: Vec<SyntheticDefinition>,
}

A Move module

Fields
loc: Loc

The location of this module

identifier: ModuleIdent

name and address of the module

friends: Vec<ModuleIdent>

the module’s friends

imports: Vec<ImportDefinition>

the module’s dependencies

explicit_dependency_declarations: Vec<ModuleDependency>

Explicit declaration of dependencies. If not provided, will be inferred based on given dependencies to the IR compiler

structs: Vec<StructDefinition>

the structs (including resources) that the module defines

constants: Vec<Constant>

the constants that the script defines. Only a utility, the identifiers are not carried into the Move bytecode

functions: Vec<(FunctionName, Function)>

the procedure that the module defines

synthetics: Vec<SyntheticDefinition>

the synthetic, specification variables the module defines.

Implementations

impl ModuleDefinition

pub fn new(
loc: Loc,
identifier: ModuleIdent,
friends: Vec<ModuleIdent>,
imports: Vec<ImportDefinition>,
explicit_dependency_declarations: Vec<ModuleDependency>,
structs: Vec<StructDefinition>,
constants: Vec<Constant>,
functions: Vec<(FunctionName, Function)>,
synthetics: Vec<SyntheticDefinition>
) -> Self

Creates a new ModuleDefinition from its string name, dependencies, structs+resources, and procedures Does not verify the correctness of any internal properties of its elements

pub fn get_external_deps(&self) -> Vec<ModuleId>

Return a vector of ModuleId for the external dependencies.