Enum move_ir_types::ast::Exp_

pub enum Exp_ {
    Dereference(Box<Exp>),
    UnaryExp(UnaryOp, Box<Exp>),
    BinopExp(Box<Exp>, BinOp, Box<Exp>),
    Value(CopyableVal),
    Pack(StructName, Vec<Type>, ExpFields),
    Borrow {
        is_mutable: bool,
        exp: Box<Exp>,
        field: FieldIdent,
    },
    Move(Var),
    Copy(Var),
    BorrowLocal(bool, Var),
    FunctionCall(FunctionCall, Box<Exp>),
    ExprList(Vec<Exp>),
}

Enum for all expressions

Variants

Dereference(Box<Exp>)

*e

Tuple Fields of Dereference

0: Box<Exp>

UnaryExp(UnaryOp, Box<Exp>)

op e

Tuple Fields of UnaryExp

0: UnaryOp1: Box<Exp>

BinopExp(Box<Exp>, BinOp, Box<Exp>)

e_1 op e_2

Tuple Fields of BinopExp

0: Box<Exp>1: BinOp2: Box<Exp>

Value(CopyableVal)

Wrapper to lift CopyableVal into Exp v

Tuple Fields of Value

0: CopyableVal

Pack(StructName, Vec<Type>, ExpFields)

Takes the given field values and instantiates the struct Returns a fresh StructInstance whose type and kind (resource or otherwise) as the current struct class (i.e., the class of the method we’re currently executing). n { f_1: e_1, ... , f_j: e_j }

Tuple Fields of Pack

0: StructName1: Vec<Type>2: ExpFields

Borrow

&e.f, &mut e.f

Fields of Borrow

is_mutable: bool

mutable or not

exp: Box<Exp>

the expression containing the reference

field: FieldIdent

the field being borrowed

Move(Var)

move(x)

Tuple Fields of Move

0: Var

Copy(Var)

copy(x)

Tuple Fields of Copy

0: Var

BorrowLocal(bool, Var)

&x or &mut x

Tuple Fields of BorrowLocal

0: bool1: Var

FunctionCall(FunctionCall, Box<Exp>)

f(e) or f(e_1, e_2, ..., e_j)

Tuple Fields of FunctionCall

0: FunctionCall1: Box<Exp>

ExprList(Vec<Exp>)

(e_1, e_2, e_3, …, e_j)

Tuple Fields of ExprList

0: Vec<Exp>

Implementations

impl Exp_

pub fn address(addr: AccountAddress) -> Exp

Creates a new address Exp with no location information

pub fn value(b: CopyableVal_) -> Exp

Creates a new value Exp with no location information

pub fn u64(i: u64) -> Exp

Creates a new u64 Exp with no location information

pub fn bool(b: bool) -> Exp

Creates a new bool Exp with no location information

pub fn byte_array(buf: Vec<u8>) -> Exp

Creates a new bytearray Exp with no location information

pub fn instantiate(n: StructName, tys: Vec<Type>, s: ExpFields) -> Exp

Creates a new pack/struct-instantiation Exp with no location information

pub fn binop(lhs: Exp, op: BinOp, rhs: Exp) -> Exp

Creates a new binary operator Exp with no location information

pub fn add(lhs: Exp, rhs: Exp) -> Exp

Creates a new e+e Exp with no location information

pub fn sub(lhs: Exp, rhs: Exp) -> Exp

Creates a new e-e Exp with no location information

pub fn dereference(e: Exp) -> Exp

Creates a new *e Exp with no location information

pub fn borrow(is_mutable: bool, exp: Box<Exp>, field: FieldIdent) -> Exp

Creates a new borrow field Exp with no location information

pub fn copy(v: Var) -> Exp

Creates a new copy-local Exp with no location information

pub fn move_(v: Var) -> Exp

Creates a new move-local Exp with no location information

pub fn function_call(f: FunctionCall, e: Exp) -> Exp

Creates a new function call Exp with no location information

pub fn expr_list(exps: Vec<Exp>) -> Exp