Module move_ir_types::location

Structs

Loc

The Loc struct is used to define a location in a file; where the file is considered to be a vector of bytes, and the range for a given Loc is defined by start and end index into that byte vector

Spanned

Functions

sp

Function used to have nearly tuple-like syntax for creating a Spanned

Type Definitions

ByteIndex

An index into a file. Much like the codespan crate, a u32 is used here to for space efficiency. However, this assumes no file is larger than 4GB, so this might become a usize in the future if the space concerns turn out to not be an issue.